requirements.txt
# Alpaca Trading API
alpaca-py>=0.21.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Environment variables
python-dotenv>=1.0.0

# Timezone support (Python 3.9+)
# zoneinfo is built-in for Python 3.9+

# Optional: News/Sentiment APIs
# newsapi-python>=0.2.7  # Uncomment if using NewsAPI
# tweepy>=4.14.0         # Uncomment if using Twitter

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# Optional: Visualization for backtesting
# matplotlib>=3.7.0
# plotly>=5.15.0
